chore: structure files
