+++ core/src/constants/index.ts
@@ -6,3 +6,5 @@ export const packagePrefix = 'oreo';
 export const cssVariablePrefix = `--${packagePrefix}`;
 
 export const defaultStandardFontSize = 16;
+
+export type * from './index.types';
