+++ core/src/hooks/use-disclosure/index.ts
@@ -1,1 +1,2 @@
 export { useDisclosure } from './use-disclosure';
+export type { UseDisclosureCallbackType } from './use-disclosure.type';
