+++ core/src/hooks/use-keydown/index.ts
@@ -1,1 +1,2 @@
 export { useKeydown } from './use-keydown';
+export type { UseKeydownkOptionsType } from './use-keydown';
