+++ core/src/hooks/use-outside-element-click/index.ts
@@ -1,1 +1,2 @@
 export { useOutsideElementClick } from './use-outside-element-click';
+export type { UseOutsideElementClickOptionsType } from './use-outside-element-click';
