+++ core/src/hooks/use-timeout/index.ts
@@ -1,1 +1,2 @@
 export { useTimeout } from './use-timeout';
+export type { UseTimeoutOptionsType } from './use-timeout.type';
