+++ core/src/styled/mixins/mixin.ts
@@ -1,0 +1,26 @@
+import { isPackageNative } from '../../helpers/base';
+import { getDimensionsUnitHandler } from '../../helpers/layout';
+import { breakpointsObject as breakpoints } from '../../theme/utilities/breakpoints';
+import type {
+  MediaStylesTypes,
+  MediaStylesSizeTypes,
+  MediaStylesStyleTypes,
+} from './mixin.type';
+
+export const mediaQueryStyle = (
+  key: MediaStylesSizeTypes,
+  style: MediaStylesStyleTypes,
+  options?: MediaStylesTypes
+) => {
+  const packageType = options?.packageType || 'web';
+
+  const breakpoint = isPackageNative(packageType)
+    ? breakpoints[key]
+    : getDimensionsUnitHandler(breakpoints[key]).rem;
+
+  return `
+    @media screen and (min-width: ${breakpoint}) {
+      ${style}
+    }
+  `;
+};
