+++ core/src/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './context/index';
+export type { ProviderProps } from './context/index.types';
