+++ core/styled.d.ts
@@ -1,0 +1,7 @@
+import 'styled-components';
+import type { ThemeType } from './src/theme';
+
+// @extends
+declare module 'styled-components' {
+  export interface DefaultTheme extends ThemeType {}
+}
