+++ core/tsconfig.json
@@ -1,4 +1,10 @@
 {
   "extends": "../tsconfig.json",
-  "include": ["src"]
+  "compilerOptions": {
+    "paths": {
+      "@/*": ["./src/*"],
+    },
+  },
+  "include": ["src"],
+  
 }
